% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AutoPipe_tSNE.R
\name{AutoPipe_tSNE}
\alias{AutoPipe_tSNE}
\title{Implemented t-distributed stochastic neighbor embedding}
\usage{
AutoPipe_tSNE(me,perplexity=30,max_iter=500)
}
\arguments{
\item{me}{The path of the expression table}

\item{perplexity}{numeric; Perplexity parameter}

\item{max_iter}{integer; Number of iterations (default: 1000)}
}
\description{
This function is used to upload a table into R for further use in the AutoPipe
}
