% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cons_clust.R
\name{cons_clust}
\alias{cons_clust}
\title{A function to plot do a Consensus clustering to validate the results}
\usage{
cons_clust(data,max_clust,TOPgenes)
}
\arguments{
\item{data}{this is the data for the ConsensusClusterPlus}

\item{max_clust}{the max number of clusters that should be evaluated.}

\item{TOPgenes}{the number of the top genes to choose for the clustering}
}
\value{
plots a plot with all the heatmaps from the ConsensusClusterPlus for the number ofd clusters 2 to max_clust
 the same return value as the COnsensusClusterPlus
}
\description{
this function  calls the ConsensusClusterPlus function with thedaraset and plots a plot
with the heatmaps of the clustering for each number of clusters from 2 to max_clust
}
\examples{

data(rna)
cons_clust(rna,5,TOPgenes=50)

}
