% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Groups_Sup.R
\name{Groups_Sup}
\alias{Groups_Sup}
\title{cluster the samples}
\usage{
Groups_Sup(me_TOP, me, number_of_k,TRw)
}
\arguments{
\item{me_TOP}{the matrix with the n top genes, usually the  from
output of the function TopPAM}

\item{me}{the original expression matrix. (with genes in rows and samples in columns).}

\item{number_of_k}{the number of clusters}

\item{TRw}{threshold for the elemenation of the samples with a Silhouette width lower than TRw.
Default value is -1.}
}
\description{
This function clusters the samples into x clusters.
}
\examples{

## load data
library(org.Hs.eg.db)
data(rna)
me_x=rna
res<-AutoPipe::TopPAM(me_x,max_clusters = 8, TOP=100)
me_TOP=res[[1]]
number_of_k=res[[3]]
File_genes=Groups_Sup(me_TOP, me=me_x, number_of_k,TRw=-1)
groups_men=File_genes[[2]]
me_x=File_genes[[1]]

}
