% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_expression_file.R
\name{read_expression_file}
\alias{read_expression_file}
\title{Input Expression File}
\usage{
read_expression_file(file, format = "csv", sep=";",gene_name="SYMBOL", Trans=FALSE)
}
\arguments{
\item{file}{The path of the expression table}

\item{format}{The format of the table "csv" or "txt"}

\item{sep}{The seperator of the input table}

\item{gene_name}{Genes are given in "SYMBOL" or "ENTREZID"}

\item{Trans}{Need Matrix Transpose TRUE or FALSE}
}
\value{
A data.frame with a gene expression matrix
}
\description{
This function is used to upload a table into R for further use in the AutoPipe
}
