% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TopPAM.R
\name{TopPAM}
\alias{TopPAM}
\title{Compute Top genes}
\usage{
TopPAM(me, max_clusters=15,TOP=1000)
}
\arguments{
\item{me}{a matrix with genes in rows and samples in columns}

\item{max_clusters}{max. number of clusters to check}

\item{TOP}{the number of genes to take}
}
\value{
a list of 1. A matrix with the top genes
2. A list of means of the Silhouette width for each number of clusters. 3. The optimal number of clusters
}
\description{
This function computes the n=TOP genes and the the best number of clusters
}
\examples{
##load the org.Hs.eg Library
library(org.Hs.eg.db)
#' ## load data
data(rna)
me_x=rna
res<-AutoPipe::TopPAM(me_x,max_clusters = 8, TOP=1000)
me_TOP=res[[1]]
number_of_k=res[[3]]

}
