% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UnSuperClassifier.R
\name{UnSuperClassifier}
\alias{UnSuperClassifier}
\title{Unsupervised Clustering}
\usage{
UnSuperClassifier(data,clinical_data=NULL)
}
\arguments{
\item{data}{the data for the clustering. Data should be in the following format: samples in columns and
the genes in the rows (colnames and rownames accordingly). The rownames should be Entrez ID in order to
plot a gene set enrichment analysis.}

\item{clinical_data}{the clinical data provided by the user to plot under the heatmap. it will be
plotted only if show_clin is TRUE. Default value is NULL. see details for format.
@details sample data should be a data.frame with the sample names
as rownames and the clinical triats as columns.
 each trait must be a numeric variable.}
}
\description{
A function for unsupervised Clustering of the data
}
