% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{y}
\alias{y}
\title{rna egene expression of 10 Tumors}
\format{A data frame with 10798 rows and 10 variables:
\describe{
  \item{X1._.1}{sample X1._.6, }
  \item{X1._.2}{sample X1._.2, }
  \item{X1._.3}{sample X1._.3, }
  \item{X1._.4}{sample X1._.4, }
  \item{X1._.5}{sample X1._.5, }
  \item{X2._.6}{sample X2._.6, }
  \item{X2._.7}{sample X2._.7, }
  \item{X2._.8}{sample X2._.8, }
  \item{X2._.9}{sample X2._.9, }
  \item{X2._.1}{sample X2._.1, }
  ...
}}
\usage{
y
}
\description{
A dataset containing the gene expression data od 10  tumors
}
\keyword{datasets}
